Code/Server/Django_Web_REST_Server/requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
crypto==1.4.1
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.42.0
django-filter==2.2.0
django-filters==0.2.1
django-oauth-toolkit==1.3.2
django-session-timeout==0.1.0
django-tables2==2.3.1
djangorestframework==3.11.0
djangorestframework-oauth==1.1.0
ecdsa==0.15
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Naked==0.1.31
oauthlib==3.1.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycparser==2.20
pylint==2.5.3
python-jose==3.1.0
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
shellescape==3.8.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.8
wrapt==1.12.1
